+++ kork-core/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,5 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.netflix.spinnaker.kork.PlatformComponents,com.netflix.spinnaker.kork.secrets.SecretConfiguration
+  com.netflix.spinnaker.kork.PlatformComponents
 
 
 org.springframework.context.ApplicationListener=\
