+++ kork-secrets/src/main/java/com/netflix/spinnaker/kork/secrets/EncryptedSecret.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.config.secrets;
+package com.netflix.spinnaker.kork.secrets;
 
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
