+++ kork-secrets/src/test/java/com/netflix/spinnaker/kork/secrets/EncryptedSecretTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.config.secrets;
+package com.netflix.spinnaker.kork.secrets;
 
 import org.junit.Rule;
 import org.junit.Test;
